Time: 0 ms (100%), Space: 11.1 MB (19.33%) - LeetHub

+++ 0013-roman-to-integer/0013-roman-to-integer.c
@@ -74,7 +74,7 @@ int romanToInt(char* s) {
             if (tmp != C)
                 i++;
         }
-        result = result + tmp;
+        result = result + (long long)tmp;
     }
-    return (result);
+    return ((int)result);
 }
